requirements.txt
py-cord>=2.4.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
aiomysql>=0.2.0
sqlalchemy>=2.0.0
pynacl>=1.5.0
